normals: avoid disassociation with xyz positions
Prior to this change, the normal data is loaded into a separate point
cloud object from a separate file without xyz coordinates, relying on
an implicit assumption that the normal vectors maintain a 1:1
correspondence with the xyz position data of file1(-a).

However, this is neither robust to the sorted order of the points in
file1 and the normal file being different, nor to operations that
modify the order of xyz positions after loading the data (such as
duplicate point removal).

This commit ensures that xyz data is loaded along with normal vectors
and is maintained in the same object.

Additionally, in the case when the file1(-a) and the normal file(-n)
are the same, the file is parsed only once.

No special regard is given to the behaviour of normals during duplicate
point removal.  The normal vector of the first point is retained.