memory: avoid expensive per-point structure in duplicate point removal
The need to maintain a 40 byte per-point structure in an associative
array in order to tally identical point attributes and control the
retention of duplicate points can be avoided by first loading all
points, sorting them, and then processing duplicates sequentially,
thereby reducing the amount of state required.

A side-effect of this process is that points are always in the sorted
order for subsequent processing stages regardless of input order.